pyproject.toml
[project]
name = "spotify-data-analysis-refactor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pychalk>=2.0.1",
    "rapidfuzz>=3.14.3",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "squarify>=0.4.4",
]

[tool.ruff]
line-length = 120
